+++ R/MaestroPipeline.R
@@ -53,7 +53,7 @@ MaestroPipeline <- R6::R6Class(
       private$outputs <- outputs
       private$priority <- priority
       private$flags <- flags
-      private$run_if <- run_if
+      private$run_if <- if (!is.null(run_if) && trimws(run_if) == "") NULL else run_if
 
       if (is.null(inputs)) {
 
@@ -195,10 +195,6 @@ MaestroPipeline <- R6::R6Class(
       script_path <- private$script_path
       log_level <- private$log_level
 
-      if (!quiet) {
-        cli::cli_progress_step("{cli_prepend}{cli::col_blue(pipe_name)}")
-      }
-
       if (log_to_console) {
         logger_fun <- logger::appender_tee
       } else {
@@ -229,18 +225,27 @@ MaestroPipeline <- R6::R6Class(
 
       do_run <- TRUE
       if (!is.null(private$run_if)) {
-        cond <- tryCatch({
+
+        if (!quiet) {
+          cli::cli_progress_step("? {cli::col_blue(pipe_name)}")
+        }
+
+        cond <- withCallingHandlers(
           eval_code_str(
             private$run_if,
             vars = resources,
             inherit = maestro_context
-          )
-        }, error = function(e) {
-          cli::cli_abort("maestroRunIf condition resulted in an error")
-        })
+          ),
+          error = private$cond_error_handler,
+          warning = private$cond_warning_handler,
+          message = private$cond_message_handler
+        )
 
         if (!rlang::is_scalar_logical(cond)) {
-          cli::cli_abort("maestroRunIf condition did not result in a single TRUE/FALSE.")
+          withCallingHandlers(
+            stop(glue::glue("`{private$run_if}` did not return a single boolean."), call. = FALSE),
+            error = private$cond_error_handler
+          )
         }
 
         do_run <- cond
@@ -249,6 +254,11 @@ MaestroPipeline <- R6::R6Class(
       if (!do_run) return(invisible())
 
       private$run_time_start <- lubridate::now()
+
+      if (!quiet) {
+        cli::cli_progress_step("{cli_prepend}{cli::col_blue(pipe_name)}")
+      }
+
       private$status <- "Success"
 
       results <- withCallingHandlers(
@@ -491,6 +501,29 @@ MaestroPipeline <- R6::R6Class(
     },
 
     message_handler = function(m) {
+      message_log <- logger::log_info(conditionMessage(m), namespace = private$pipe_name)
+      private$messages <- c(private$messages, m$message)
+      invokeRestart("muffleMessage")
+    },
+
+    # Handlers for conditionals
+    cond_error_handler = function(e) {
+      e$message <- paste("Error evaluating condition:", e$message)
+      private$errors <- e
+      private$status <- "Error"
+      logger::log_error("Error evaluating condition: {conditionMessage(e)}", namespace = private$pipe_name)
+      private$run_time_end <- lubridate::now()
+    },
+
+    cond_warning_handler = function(w) {
+      warning_log <- logger::log_warn("Warning evaluating condition: {conditionMessage(w)}", namespace = private$pipe_name)
+      w$message <- paste("Warned while evaluating condition:", w$message)
+      private$warnings <- c(private$warnings, w$message)
+      private$status <- "Warning"
+      invokeRestart("muffleWarning")
+    },
+
+    cond_message_handler = function(m) {
       message_log <- logger::log_info(conditionMessage(m), namespace = private$pipe_name)
       private$messages <- c(private$messages, m$message)
       invokeRestart("muffleMessage")