+++ NEWS.md
@@ -1,3 +1,15 @@
+# maestro 0.7.0.9000
+
+### New features
+
+- Conditional pipelines: A new `maestroRunIf` tag that takes an R expression returning a boolean and conditionally runs a pipeline. These can use inputs from upstream DAG pipelines, values from the orchestrator via `run_schedule(..., resources = list())`, or any arbitrary TRUE/FALSE expression.
+
+### Minor changes
+
+- `maestro::invoke()` gains argument `log_to_console` allowing it to print messages, warnings, and errors to the console.
+
+- `last_run_errors()` and friends now return named lists where the name corresponds to the pipe name.
+
 # maestro 0.6.3
 
 ### Bug fixes